Changed the role paradigm to add pix and set the path to cards
Changed how the checkout works. Now it starts by choosing the payment method. If it's pix we are testing if it works on prod that why I'm commiting right now, it won't work with a test secret key.

If it's a credit card it follow its normal path that I already created, only adjusted to work with this new paradigm